#43 refactor: DiaryDetailResDto 제거
