+++ src/main/java/com/dasibom/practice/dto/DiaryDto.java
@@ -1,6 +1,8 @@
 package com.dasibom.practice.dto;
 
 import com.dasibom.practice.domain.Diary;
+import com.dasibom.practice.domain.DiaryImage;
+import com.dasibom.practice.domain.DiaryStamp;
 import com.dasibom.practice.domain.Pet;
 import com.dasibom.practice.domain.Stamp;
 import java.time.LocalDateTime;
@@ -81,5 +83,38 @@ public class DiaryDto {
             this.createdAt = diary.getCreatedAt().toString();
         }
     }
+
+    @Getter
+    @AllArgsConstructor
+    public static class DetailResponse {
+        private final String title;
+        private final String content;
+        private final String petName;
+        private final List<String> imgUrls = new ArrayList<>();
+        private final List<String> stampTypes = new ArrayList<>();
+        private final LocalDateTime createdAt;
+
+        public DetailResponse(Diary entity) {
+            this.title = entity.getTitle();
+            this.content = entity.getContent();
+            this.petName = entity.getPet().getPetName();
+            this.createdAt = entity.getCreatedAt();
+            initImgUrls(entity);
+            initStampTypes(entity);
+        }
+
+        private void initImgUrls(Diary entity) {
+            for (DiaryImage image : entity.getImages()) {
+                this.imgUrls.add(image.getImgUrl());
+            }
+        }
+
+        private void initStampTypes(Diary entity) {
+            for (DiaryStamp diaryStamp : entity.getDiaryStamps()) {
+                this.stampTypes.add(diaryStamp.getStamp().getStampType().toString());
+            }
+        }
+    }
+
 
 }
