+++ src/main/java/com/dasibom/practice/repository/custom/CustomDiaryRepository.java
@@ -1,7 +1,6 @@
 package com.dasibom.practice.repository.custom;
 
 import com.dasibom.practice.condition.DiaryReadCondition;
-import com.dasibom.practice.dto.DiaryDetailResDto;
 import com.dasibom.practice.dto.DiaryDto;
 import java.util.List;
 import org.springframework.data.domain.Pageable;
@@ -11,6 +10,6 @@ public interface CustomDiaryRepository {
 
     Slice<DiaryDto.SimpleResponse> getDiaryBriefInfoScroll(Long cursorId, DiaryReadCondition condition, Pageable pageable);
 
-    List<DiaryDetailResDto> getDiaryDetailList(DiaryReadCondition condition);
+    List<DiaryDto.DetailResponse> getDiaryDetailList(DiaryReadCondition condition);
 
 }
